Move some helpers to RefinanceTokenModal
